Remove direct inclusion of codecs from examples, add Pako as dependency
