+++ package.json
@@ -81,6 +81,7 @@
     "worker-loader": "^2.0.0"
   },
   "dependencies": {
-    "dicom-parser": "^1.8.3"
+    "dicom-parser": "^1.8.3",
+    "pako": "^1.0.10"
   }
 }
