+++ src/webWorker/index.worker.js
@@ -3,7 +3,8 @@ import decodeTask from './decodeTask/decodeTask.js';
 import { default as version } from '../version.js';
 
 // TODO: Not sure if this is necessary, but we probably need it for deflate in dicomParser
-import pako from '../../codecs/pako.min.js';
+/* eslint-disable-next-line import/extensions */
+import 'pako';
 
 // register our task
 registerTaskHandler(decodeTask);
